style: Add mock html function which provides syntaxt highlighting
Since we have quite some markup generated in plain `.js` files, we run
in the problem that we don't have syntaxt highlighting. This lib
function in combination with the LitHtml extension for VS Code solves
this problem

This is implemented by creating a mock tagged template literal function.
The LitHtml extension for VSCode generates syntaxt highlighting for html
when it finds a tagged template literal with the tag of html. The lib
function is a mock function