refactor(sitemap.view): Update error message in web app
The previous error message explained that the /sitemap.xml could not be
found. But we never look at /sitemap.xml when looking for sitemaps. The
api is using /robots.txt to search for sitemap references.

This error message explains that. Otherwise, a sitemap.xml can be
present on the server, but could not be found by us because we are not
looking for it directly